Animation move Player to Controller
