+++ src/util/const.js
@@ -1,5 +1,7 @@
 const FRAME_SIZE = 20;
 const TIMELINE_HEIGHT = 28;
 const INIT_MAX_FRAME = 20;
+const FRAME_SEC = 33; // 30 fps
 
-export { FRAME_SIZE, TIMELINE_HEIGHT, INIT_MAX_FRAME };
+
+export { FRAME_SIZE, TIMELINE_HEIGHT, INIT_MAX_FRAME, FRAME_SEC };
